Add initial set of unit tests for OpensslCertChainValidator
Motivation:

Issue #116 affected dCache sites, requiring work-arounds.  To facilitate
resolving this issue, it is helpful to recreate the problem as a
unit-test.

Modification:

Introduce a generic framework for testing OpensslCertChainValidator,
based on BDD and various builder patterns.  The emphasis is on making
the unit-tests readable by hiding low-level details while avoiding any
test fixtures.

Include a couple of tests that pass (which show that the unit tests work
as expected) and a unit test that demonstrates the problem highlighted
in issue #116.

Result:

It is now possible to test whether CaNL suffers from issue #116.