Fix order of certificates
Motivation:

The certificate chain (as supplied by the server) is not arbitrary, but
should start with the EEC.

The server certificate chain that triggered issue #116 does not follow
this prescribed order.  However, the real-world example also doesn't
follow the order used in the unit-test.  Moreover, the underlying
problem is present with correctly ordered certificates.

Therefore, in order to provide a minimal example that demonstrates the
problem, we should use the correct certificate ordering.

Modification:

Update order to place the EEC first, intermediate CAs next and finally
the root CA.

Result:

No change in behaviour; however, the certificate order more closely
resembles correct behaviour.